Update 1
Added base files and sounds.  Added genSequence function to create a random sequence of numbers up to a given input length and save them into a string.  Added a user move function and buttons.  The functions take button input and save them into a sequence.  Created a verifySeq string to verify if the user sequence string is the same as the generated sequence.