Update 9
Added grid elements to the css and design.  Added an error sound and input error into the code when a user clicks a wrong sequence button.  Added functionality to change button background colors when selected as part of the sequence.  Added functionality to change button background color when a user clicks a button.  Added hard mode that will reset the game completely if a user clicks an incorrect button in the sequence.